pyproject.toml
[project]
name = "geography-auto-grading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.12.0",
    "google-generativeai>=0.8.5",
    "groq>=0.11.0",
    "langchain-community>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "plotly>=5.17.0",
    "psutil>=5.9.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
